[APPSRE-1776] Fix url rewrite
The github api url is replaced by the github mirror url, but we can't
always probe the service url from within the service, specially when
we run the service without SSL and there's a Route facing the word which
implements SSL for us.

This fix allow us to specify the actual URL for the service, while still
trying our best to probe when the URL is not specified.

Signed-off-by: Amador Pahim <apahim@redhat.com>