Specific metrics for rate limited state
This feature was originally introduced hooking into the offline
flow, but that has the side effect of making hard to diferentiate
those from the real offline mode.

This patch removes the need for using the offline flow for serving
requests from cache when we are being rate limited. With that,
we have now a new metric cache="RATE_LIMITED_MISS/HIT" so we can generate
independent graphs.

Signed-off-by: Amador Pahim <apahim@redhat.com>