Allow to increase sleep time in github status check
We are currently hammering the status endpoint and we're getting rate
limited. Every time we get rate limited, github mirror thinks it has to
go offline, so this tries to mitigate it.

We cannot go crazy with this approach as the same sleep time is used to
enter and to exit from the offline mode.

Related to APPSRE-8243

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>