Redistribute latency buckets
There are very little requests under 0.05s (0.02% in the last 28 days),
so it doesn't make sense to keep all the granularity there. It is
equally non important to keep the granularity over 5s, since there are
almost no differences between the number of requests under 7.5s and 10s.
We have added granularity in the area where we have the most of our
requests (0.25s) as that can help us to define more precisely our SLOs.

APPSRE-9880

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>